# docs/swagger.yaml
definitions:
...
  models.Email:
    properties:
      contact_id:
        type: integer
      email:
        type: string
      id:
        type: integer
      is_primary:
        type: boolean
      label_type_id:
        type: integer
      type_label:
        type: string
    type: object
  models.EmailJSONPatch:
    properties:
      contact_id:
        example: 4
        type: integer
      id:
        example: 1
        type: integer
      is_primary:
        example: false
        type: boolean
      label_type_id:
        example: 42
        type: integer
      phone:
        example: support@kindredcard.com
        type: string
    type: object
...
  models.Phone:
    properties:
      contact_id:
        type: integer
      id:
        type: integer
      is_primary:
        type: boolean
      label_type_id:
        type: integer
      last_formatted_at:
        type: string
      phone:
        type: string
      type_label:
        type: string
    type: object
  models.PhoneJSONPatch:
    properties:
      contact_id:
        example: 4
        type: integer
      id:
        example: 1
        type: integer
      is_primary:
        example: false
        type: boolean
      label_type_id:
        example: 42
        type: integer
      phone:
        example: (555)122-4121
        type: string
    type: object
...
info:
  contact:
    name: API Support
    url: https://github.com/steveredden/KindredCard/issues
  description: Personal CRM for managing contacts, relationships, and important dates
  license:
    name: AGPLv3
    url: https://www.gnu.org/licenses/agpl-3.0.txt
  title: KindredCard API
  version: "1.0"
paths:
...
  /api/v1/contacts/{cid}/emails:
    post:
      consumes:
      - application/json
      description: Associates an email with a contact using HTTP POST
      parameters:
      - description: Contact ID
        in: path
        name: cid
        required: true
        type: integer
      - description: Email fields
        in: body
        name: url
        required: true
        schema:
          $ref: '#/definitions/models.Email'
      produces:
      - application/json
      responses:
        "200":
          description: 'created: newID'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request body or contact ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Contact not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiTokenAuth: []
      summary: Creates an email associated with a contact
      tags:
      - contacts
  /api/v1/contacts/{cid}/emails/{eid}:
    delete:
      consumes:
      - application/json
      description: Associates an email with a contact using HTTP DELETE
      parameters:
      - description: Contact ID
        in: path
        name: cid
        required: true
        type: integer
      - description: Email ID
        in: path
        name: eid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: deleted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request body or contact ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Contact not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiTokenAuth: []
      summary: Removes an email associated with a contact
      tags:
      - contacts
...
  /api/v1/contacts/{cid}/phone:
    post:
      consumes:
      - application/json
      description: Associates a Phone with a contact using HTTP POST
      parameters:
      - description: Contact ID
        in: path
        name: cid
        required: true
        type: integer
      - description: Phone fields
        in: body
        name: phone
        required: true
        schema:
          $ref: '#/definitions/models.Phone'
      produces:
      - application/json
      responses:
        "200":
          description: 'created: newID'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request body or contact ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Contact not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiTokenAuth: []
      summary: Creates a website / Phone associated with a contact
      tags:
      - contacts
  /api/v1/contacts/{cid}/phone/{pid}:
    delete:
      consumes:
      - application/json
      description: Associates a Phone with a contact using HTTP POST
      parameters:
      - description: Contact ID
        in: path
        name: cid
        required: true
        type: integer
      - description: Phone ID
        in: path
        name: pid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: deleted
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request body or contact ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Contact not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiTokenAuth: []
      summary: Creates a website / Phone associated with a contact
      tags:
      - contacts
...
  /api/v1/emails/{eid}:
    patch:
      consumes:
      - application/json
      description: Update specific fields of a email using HTTP PATCH. Only provided
        fields will be updated.
      parameters:
      - description: Email ID
        in: path
        name: eid
        required: true
        type: integer
      - description: Email fields to update
        in: body
        name: contact
        required: true
        schema:
          $ref: '#/definitions/models.EmailJSONPatch'
      produces:
      - application/json
      responses:
        "200":
          description: Updated email
          schema:
            items:
              $ref: '#/definitions/models.Email'
            type: array
        "400":
          description: Invalid request body or contact ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Contact not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiTokenAuth: []
      summary: Update a email number
      tags:
      - contacts
...
  /api/v1/phones/{pid}:
    patch:
      consumes:
      - application/json
      description: Update specific fields of a phone using HTTP PATCH. Only provided
        fields will be updated.
      parameters:
      - description: Phone ID
        in: path
        name: pid
        required: true
        type: integer
      - description: Phone fields to update
        in: body
        name: contact
        required: true
        schema:
          $ref: '#/definitions/models.PhoneJSONPatch'
      produces:
      - application/json
      responses:
        "200":
          description: Updated phone
          schema:
            items:
              $ref: '#/definitions/models.Phone'
            type: array
        "400":
          description: Invalid request body or contact ID
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Contact not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiTokenAuth: []
      summary: Update a phone number
      tags:
      - contacts
...
securityDefinitions:
  ApiTokenAuth:
    description: API token for authentication
    in: header
    name: session
    type: apiKey
swagger: "2.0"